square

+++ include/LSGFW/lsgfw.h
@@ -58,8 +58,8 @@ typedef unsigned 	char		u8_t;
 typedef signed		char		i8_t;
 typedef unsigned 	short		u16_t;
 typedef signed		short		i16_t;
-typedef unsigned 	long		u32_t;
-typedef signed		long		i32_t;
+typedef unsigned 	int		u32_t;
+typedef signed		int		i32_t;
 typedef unsigned	long long	u64_t;
 typedef signed		long long	i64_t;
